% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.backup_operations.R
\name{delete_backup_selection}
\alias{delete_backup_selection}
\title{Deletes the resource selection associated with a backup plan that is specified by the SelectionId}
\usage{
delete_backup_selection(BackupPlanId, SelectionId)
}
\arguments{
\item{BackupPlanId}{[required] Uniquely identifies a backup plan.}

\item{SelectionId}{[required] Uniquely identifies the body of a request to assign a set of resources to a backup plan.}
}
\description{
Deletes the resource selection associated with a backup plan that is specified by the \code{SelectionId}.
}
\section{Accepted Parameters}{
\preformatted{delete_backup_selection(
  BackupPlanId = "string",
  SelectionId = "string"
)
}
}

