% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.backup_operations.R
\name{delete_backup_vault}
\alias{delete_backup_vault}
\title{Deletes the backup vault identified by its name}
\usage{
delete_backup_vault(BackupVaultName)
}
\arguments{
\item{BackupVaultName}{[required] The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and theAWS Region where they are created. They consist of lowercase letters, numbers, and hyphens.}
}
\description{
Deletes the backup vault identified by its name. A vault can be deleted only if it is empty.
}
\section{Accepted Parameters}{
\preformatted{delete_backup_vault(
  BackupVaultName = "string"
)
}
}

