% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.backup_operations.R
\name{delete_backup_vault_access_policy}
\alias{delete_backup_vault_access_policy}
\title{Deletes the policy document that manages permissions on a backup vault}
\usage{
delete_backup_vault_access_policy(BackupVaultName)
}
\arguments{
\item{BackupVaultName}{[required] The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the AWS Region where they are created. They consist of lowercase letters, numbers, and hyphens.}
}
\description{
Deletes the policy document that manages permissions on a backup vault.
}
\section{Accepted Parameters}{
\preformatted{delete_backup_vault_access_policy(
  BackupVaultName = "string"
)
}
}

