% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.backup_operations.R
\name{describe_backup_job}
\alias{describe_backup_job}
\title{Returns metadata associated with creating a backup of a resource}
\usage{
describe_backup_job(BackupJobId)
}
\arguments{
\item{BackupJobId}{[required] Uniquely identifies a request to AWS Backup to back up a resource.}
}
\description{
Returns metadata associated with creating a backup of a resource.
}
\section{Accepted Parameters}{
\preformatted{describe_backup_job(
  BackupJobId = "string"
)
}
}

