% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.backup_operations.R
\name{describe_backup_vault}
\alias{describe_backup_vault}
\title{Returns metadata about a backup vault specified by its name}
\usage{
describe_backup_vault(BackupVaultName)
}
\arguments{
\item{BackupVaultName}{[required] The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the AWS Region where they are created. They consist of lowercase letters, numbers, and hyphens.}
}
\description{
Returns metadata about a backup vault specified by its name.
}
\section{Accepted Parameters}{
\preformatted{describe_backup_vault(
  BackupVaultName = "string"
)
}
}

