% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.backup_operations.R
\name{describe_restore_job}
\alias{describe_restore_job}
\title{Returns metadata associated with a restore job that is specified by a job ID}
\usage{
describe_restore_job(RestoreJobId)
}
\arguments{
\item{RestoreJobId}{[required] Uniquely identifies the job that restores a recovery point.}
}
\description{
Returns metadata associated with a restore job that is specified by a job ID.
}
\section{Accepted Parameters}{
\preformatted{describe_restore_job(
  RestoreJobId = "string"
)
}
}

