% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.backup_operations.R
\name{export_backup_plan_template}
\alias{export_backup_plan_template}
\title{Returns the backup plan that is specified by the plan ID as a backup template}
\usage{
export_backup_plan_template(BackupPlanId)
}
\arguments{
\item{BackupPlanId}{[required] Uniquely identifies a backup plan.}
}
\description{
Returns the backup plan that is specified by the plan ID as a backup template.
}
\section{Accepted Parameters}{
\preformatted{export_backup_plan_template(
  BackupPlanId = "string"
)
}
}

