% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.backup_operations.R
\name{get_backup_plan}
\alias{get_backup_plan}
\title{Returns the body of a backup plan in JSON format, in addition to plan metadata}
\usage{
get_backup_plan(BackupPlanId, VersionId = NULL)
}
\arguments{
\item{BackupPlanId}{[required] Uniquely identifies a backup plan.}

\item{VersionId}{Unique, randomly generated, Unicode, UTF-8 encoded strings that are at most 1,024 bytes long. Version IDs cannot be edited.}
}
\description{
Returns the body of a backup plan in JSON format, in addition to plan metadata.
}
\section{Accepted Parameters}{
\preformatted{get_backup_plan(
  BackupPlanId = "string",
  VersionId = "string"
)
}
}

