% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.backup_operations.R
\name{get_backup_plan_from_json}
\alias{get_backup_plan_from_json}
\title{Returns a valid JSON document specifying a backup plan or an error}
\usage{
get_backup_plan_from_json(BackupPlanTemplateJson)
}
\arguments{
\item{BackupPlanTemplateJson}{[required] A customer-supplied backup plan document in JSON format.}
}
\description{
Returns a valid JSON document specifying a backup plan or an error.
}
\section{Accepted Parameters}{
\preformatted{get_backup_plan_from_json(
  BackupPlanTemplateJson = "string"
)
}
}

