% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.backup_operations.R
\name{get_backup_plan_from_template}
\alias{get_backup_plan_from_template}
\title{Returns the template specified by its templateId as a backup plan}
\usage{
get_backup_plan_from_template(BackupPlanTemplateId)
}
\arguments{
\item{BackupPlanTemplateId}{[required] Uniquely identifies a stored backup plan template.}
}
\description{
Returns the template specified by its \code{templateId} as a backup plan.
}
\section{Accepted Parameters}{
\preformatted{get_backup_plan_from_template(
  BackupPlanTemplateId = "string"
)
}
}

