% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.backup_operations.R
\name{get_backup_selection}
\alias{get_backup_selection}
\title{Returns selection metadata and a document in JSON format that specifies a list of resources that are associated with a backup plan}
\usage{
get_backup_selection(BackupPlanId, SelectionId)
}
\arguments{
\item{BackupPlanId}{[required] Uniquely identifies a backup plan.}

\item{SelectionId}{[required] Uniquely identifies the body of a request to assign a set of resources to a backup plan.}
}
\description{
Returns selection metadata and a document in JSON format that specifies a list of resources that are associated with a backup plan.
}
\section{Accepted Parameters}{
\preformatted{get_backup_selection(
  BackupPlanId = "string",
  SelectionId = "string"
)
}
}

