% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.backup_operations.R
\name{get_backup_vault_access_policy}
\alias{get_backup_vault_access_policy}
\title{Returns the access policy document that is associated with the named backup vault}
\usage{
get_backup_vault_access_policy(BackupVaultName)
}
\arguments{
\item{BackupVaultName}{[required] The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the AWS Region where they are created. They consist of lowercase letters, numbers, and hyphens.}
}
\description{
Returns the access policy document that is associated with the named backup vault.
}
\section{Accepted Parameters}{
\preformatted{get_backup_vault_access_policy(
  BackupVaultName = "string"
)
}
}

