% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.backup_operations.R
\name{get_backup_vault_notifications}
\alias{get_backup_vault_notifications}
\title{Returns event notifications for the specified backup vault}
\usage{
get_backup_vault_notifications(BackupVaultName)
}
\arguments{
\item{BackupVaultName}{[required] The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the AWS Region where they are created. They consist of lowercase letters, numbers, and hyphens.}
}
\description{
Returns event notifications for the specified backup vault.
}
\section{Accepted Parameters}{
\preformatted{get_backup_vault_notifications(
  BackupVaultName = "string"
)
}
}

