% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.backup_operations.R
\name{get_supported_resource_types}
\alias{get_supported_resource_types}
\title{Returns the AWS resource types supported by AWS Backup}
\usage{
get_supported_resource_types()
}
\description{
Returns the AWS resource types supported by AWS Backup.
}
