% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.backup_operations.R
\name{list_backup_plan_templates}
\alias{list_backup_plan_templates}
\title{Returns metadata of your saved backup plan templates, including the template ID, name, and the creation and deletion dates}
\usage{
list_backup_plan_templates(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{The next item following a partial list of returned items. For example, if a request is made to return \code{maxResults} number of items, \code{NextToken} allows you to return more items in your list starting at the location pointed to by the next token.}

\item{MaxResults}{The maximum number of items to be returned.}
}
\description{
Returns metadata of your saved backup plan templates, including the template ID, name, and the creation and deletion dates.
}
\section{Accepted Parameters}{
\preformatted{list_backup_plan_templates(
  NextToken = "string",
  MaxResults = 123
)
}
}

