% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.backup_operations.R
\name{list_backup_plans}
\alias{list_backup_plans}
\title{Returns metadata of your saved backup plans, including Amazon Resource Names (ARNs), plan IDs, creation and deletion dates, version IDs, plan names, and creator request IDs}
\usage{
list_backup_plans(NextToken = NULL, MaxResults = NULL,
  IncludeDeleted = NULL)
}
\arguments{
\item{NextToken}{The next item following a partial list of returned items. For example, if a request is made to return \code{maxResults} number of items, \code{NextToken} allows you to return more items in your list starting at the location pointed to by the next token.}

\item{MaxResults}{The maximum number of items to be returned.}

\item{IncludeDeleted}{A Boolean value with a default value of \code{FALSE} that returns deleted backup plans when set to \code{TRUE}.}
}
\description{
Returns metadata of your saved backup plans, including Amazon Resource Names (ARNs), plan IDs, creation and deletion dates, version IDs, plan names, and creator request IDs.
}
\section{Accepted Parameters}{
\preformatted{list_backup_plans(
  NextToken = "string",
  MaxResults = 123,
  IncludeDeleted = TRUE|FALSE
)
}
}

