% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.backup_operations.R
\name{list_backup_selections}
\alias{list_backup_selections}
\title{Returns an array containing metadata of the resources associated with the target backup plan}
\usage{
list_backup_selections(BackupPlanId, NextToken = NULL,
  MaxResults = NULL)
}
\arguments{
\item{BackupPlanId}{[required] Uniquely identifies a backup plan.}

\item{NextToken}{The next item following a partial list of returned items. For example, if a request is made to return \code{maxResults} number of items, \code{NextToken} allows you to return more items in your list starting at the location pointed to by the next token.}

\item{MaxResults}{The maximum number of items to be returned.}
}
\description{
Returns an array containing metadata of the resources associated with the target backup plan.
}
\section{Accepted Parameters}{
\preformatted{list_backup_selections(
  BackupPlanId = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

