% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.backup_operations.R
\name{list_backup_vaults}
\alias{list_backup_vaults}
\title{Returns a list of recovery point storage containers along with information about them}
\usage{
list_backup_vaults(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{The next item following a partial list of returned items. For example, if a request is made to return \code{maxResults} number of items, \code{NextToken} allows you to return more items in your list starting at the location pointed to by the next token.}

\item{MaxResults}{The maximum number of items to be returned.}
}
\description{
Returns a list of recovery point storage containers along with information about them.
}
\section{Accepted Parameters}{
\preformatted{list_backup_vaults(
  NextToken = "string",
  MaxResults = 123
)
}
}

