% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.backup_operations.R
\name{list_tags}
\alias{list_tags}
\title{Returns a list of key-value pairs assigned to a target recovery point, backup plan, or backup vault}
\usage{
list_tags(ResourceArn, NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{ResourceArn}{[required] An Amazon Resource Name (ARN) that uniquely identifies a resource. The format of the ARN depends on the type of resource. Valid targets for \code{ListTags} are recovery points, backup plans, and backup vaults.}

\item{NextToken}{The next item following a partial list of returned items. For example, if a request is made to return \code{maxResults} number of items, \code{NextToken} allows you to return more items in your list starting at the location pointed to by the next token.}

\item{MaxResults}{The maximum number of items to be returned.}
}
\description{
Returns a list of key-value pairs assigned to a target recovery point, backup plan, or backup vault.
}
\section{Accepted Parameters}{
\preformatted{list_tags(
  ResourceArn = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

