% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.backup_package.R
\docType{package}
\name{paws.backup-package}
\alias{paws.backup}
\alias{paws.backup-package}
\title{paws.backup: AWS Backup}
\description{
AWS Backup is a unified backup service designed to
    protect AWS services and their associated data. AWS Backup simplifies
    the creation, migration, restoration, and deletion of backups, while
    also providing reporting and auditing.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
