% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.backup_operations.R
\name{put_backup_vault_access_policy}
\alias{put_backup_vault_access_policy}
\title{Sets a resource-based policy that is used to manage access permissions on the target backup vault}
\usage{
put_backup_vault_access_policy(BackupVaultName, Policy = NULL)
}
\arguments{
\item{BackupVaultName}{[required] The name of a logical container where backups are stored. Backup vaults are identified by names that are unique to the account used to create them and the AWS Region where they are created. They consist of lowercase letters, numbers, and hyphens.}

\item{Policy}{The backup vault access policy document in JSON format.}
}
\description{
Sets a resource-based policy that is used to manage access permissions on the target backup vault. Requires a backup vault name and an access policy document in JSON format.
}
\section{Accepted Parameters}{
\preformatted{put_backup_vault_access_policy(
  BackupVaultName = "string",
  Policy = "string"
)
}
}

