% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.backup_operations.R
\name{stop_backup_job}
\alias{stop_backup_job}
\title{Attempts to cancel a job to create a one-time backup of a resource}
\usage{
stop_backup_job(BackupJobId)
}
\arguments{
\item{BackupJobId}{[required] Uniquely identifies a request to AWS Backup to back up a resource.}
}
\description{
Attempts to cancel a job to create a one-time backup of a resource.
}
\section{Accepted Parameters}{
\preformatted{stop_backup_job(
  BackupJobId = "string"
)
}
}

