% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.backup_operations.R
\name{untag_resource}
\alias{untag_resource}
\title{Removes a set of key-value pairs from a recovery point, backup plan, or backup vault identified by an Amazon Resource Name (ARN)}
\usage{
untag_resource(ResourceArn, TagKeyList)
}
\arguments{
\item{ResourceArn}{[required] An ARN that uniquely identifies a resource. The format of the ARN depends on the type of the tagged resource.}

\item{TagKeyList}{[required] A list of keys to identify which key-value tags to remove from a resource.}
}
\description{
Removes a set of key-value pairs from a recovery point, backup plan, or backup vault identified by an Amazon Resource Name (ARN)
}
\section{Accepted Parameters}{
\preformatted{untag_resource(
  ResourceArn = "string",
  TagKeyList = list(
    "string"
  )
)
}
}

